# app/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
      "title": "Diadem To-Do API Documentation",
      "version": "1.0.0",
      "description": "Documentation of the to-do API"
    },
  
  "apis": ["./src/routes/*.js"],
  "paths": {
    "/user/create": {
      "post": {
        "summary": "Create a new user account",
        "tags": ["User"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The username of the new user."
                  },
                  "password": {
                    "type": "string",
                    "description": "The password of the new user."
                  },
                  "email": {
                    "type": "string",
                    "description": "The email address of the new user."
                  },
                  "timezone": {
                    "type": "string",
                    "description": "The timezone of the new user."
                  }
                },
                "required": ["username", "password", "email", "timezone"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "timezone": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "success": {
                    "value": {
                      "message": "User successfully created",
                      "id": "5f4db09b05b724001c8d0d8c",
                      "email": "test@email.com",
                      "timezone": "Africa/Lagos"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "User not successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "error": {
                    "value": {
                      "message": "User not successfully created",
                      "error": "Username already exists"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}